security-app/src/main/java/ro/fortech/allocation/security/user/api/UserController.java
package ro.fortech.allocation.security.user.api;

import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RestController;
import ro.fortech.allocation.security.user.api.dto.UserDto;
import ro.fortech.allocation.security.user.service.UserService;

import java.util.List;

@RestController
@RequiredArgsConstructor
public class UserController implements UserApi {

    private final UserService userService;

    @Override
    public ResponseEntity<List<UserDto>> getAllUsers() {
        List<UserDto> allUsers = userService.getAllUsers();
        return ResponseEntity.ok(allUsers);
    }

    @Override

    public ResponseEntity<UserDto> getUserAtId(@PathVariable Long id) {
        UserDto user = userService.getUserAtId(id);
        return ResponseEntity.ok(user);
    }
}
